Revise Delete

+++ Shell/src/sg/edu/nus/comp/cs4218/impl/fileutils/DELETETool.java
@@ -19,8 +19,26 @@ public class DELETETool extends ATool implements IDeleteTool {
 		try{
 			if(!toDelete.exists())
 				throw new FileNotFoundException();
-			else
-				toDelete.delete();
+			else{
+				if(toDelete.isDirectory()){
+					// If it is a directory and not empty, cannot delete it 
+					if(toDelete.list().length>0){
+						System.out.println("Cannot delete the directory" + 
+					     toDelete.toString()+" since the folder is not empty");
+						return false;
+					}
+					else{
+						toDelete.delete();
+						return true;
+					}
+				}
+				// if the directory is empty, delete it
+				else{
+				// Delete the file
+					toDelete.delete();
+					return true;
+				}
+			}
 		}catch(Exception e){
 			e.printStackTrace();
 		}
